+++ policy_sentry/command/query.py
@@ -1,7 +1,8 @@
 import click
 import pprint
 import json
-from policy_sentry.shared.query import query_condition_table, query_condition_table_by_name, query_arn_table, query_arn_table_by_name
+from policy_sentry.shared.query import query_condition_table, query_condition_table_by_name, query_arn_table, \
+    query_arn_table_by_name, query_action_table
 from policy_sentry.shared.database import connect_db
 from pathlib import Path
 
@@ -52,4 +53,10 @@ def query(table, service, name):
         else:
             output = query_arn_table_by_name(db_session, service, name)
             print(json.dumps(output, indent=4))
-
+    elif table == 'action':
+        if name is None:
+            action_list = query_action_table(db_session, service)
+            for item in action_list:
+                print(item)
+    else:
+        print("Table name not valid.")